Refactoring on transformation by splitting apart into separate components
